Д/з: Авторизация и аутентификация

+++ app/Http/Controllers/Backend/LocationController.php
@@ -10,6 +10,8 @@ use App\Services\Location\LocationService;
 use App\Services\User\UserService;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Response;
+use App\Policies\Abilities;
+use Illuminate\Support\Facades\Auth;
 
 class LocationController extends Controller
 {
@@ -40,43 +42,42 @@ class LocationController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize(Abilities::VIEW_ANY, Location::class);
         return view('backend.pages.location.index', [
-            'locations' => $this->locationService->paginate(),
+            'locations' => $this->locationService->getByUser(Auth::user()),
         ]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
-        $users = [
-            // @todo Брать значение по умолчанию из конфига
-            '' => '– Please select –'
-        ];
-        foreach ($this->userService->all() as $user) {
-            $users[$user->id] = $user->name;
-        }
-        return view('backend.pages.location.create', [
-            'users' => $users,
-        ]);
+        $this->authorize(Abilities::CREATE, Location::class);
+        return view('backend.pages.location.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param  LocationStoreFormRequest  $request
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(LocationStoreFormRequest $request)
     {
-        $this->locationService->create($request->all());
+        $this->authorize(Abilities::CREATE, Location::class);
+        $data = $request->all();
+        $data['user_id'] = Auth::id();
+        $this->locationService->create($data);
         // @todo Сообщение об успешном создании записи
         return redirect(route('backend.location.index'));
     }
@@ -96,20 +97,14 @@ class LocationController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  Location  $location
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Location $location)
     {
-        $users = [
-            // @todo Брать значение по умолчанию из конфига
-            '' => '– Please select –'
-        ];
-        foreach ($this->userService->all() as $user) {
-            $users[$user->id] = $user->name;
-        }
+        $this->authorize(Abilities::UPDATE, $location);
         return view('backend.pages.location.edit', [
             'location' => $location,
-            'users' => $users,
         ]);
     }
 
@@ -118,10 +113,12 @@ class LocationController extends Controller
      *
      * @param  LocationUpdateFormRequest  $request
      * @param  Location  $location
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(LocationUpdateFormRequest $request, Location $location)
     {
+        $this->authorize(Abilities::UPDATE, $location);
         $this->locationService->update($location, $request->all());
         // @todo Сообщение об успешном обновлении записи
         return redirect(route('backend.location.index'));
@@ -131,10 +128,12 @@ class LocationController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  Location  $location
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Location $location)
     {
+        $this->authorize(Abilities::DELETE, $location);
         // @todo Промежуточная форма подтверждения
         $this->locationService->delete($location);
         // @todo Сообщение об успешном удалении записи
